2007-02-06	Joel Sherrill <joel@OARcorp.com>
	* tm09/task1.c, tm10/task1.c, tm11/task1.c, tm12/task1.c, tm13/task1.c,
	tm14/task1.c, tm22/task1.c: Address size_t/uint32_t typing issues in
	message queue tests.

+++ testsuites/tmtests/tm09/task1.c
@@ -1,6 +1,6 @@
 /*
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -103,9 +103,9 @@ void queue_test()
   uint32_t    flush_count;
   uint32_t    index;
   uint32_t    iterations;
-  long              buffer[4];
+  long        buffer[4];
   rtems_status_code status;
-  uint32_t    size;
+  size_t      size;
 
   send_loop_time    = 0;
   urgent_loop_time  = 0;
