+++ testsuites/tmtests/tm14/task1.c
@@ -1,6 +1,6 @@
 /*
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -134,7 +134,7 @@ rtems_task Low_tasks(
   rtems_task_argument argument
 )
 {
-  uint32_t   size;
+  size_t  size;
 
   (void) rtems_message_queue_receive(
            Queue_id,
